% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-linear_reg.R
\name{.linfa_linear_reg}
\alias{.linfa_linear_reg}
\title{Linear regression with linfa}
\usage{
.linfa_linear_reg(x, y)
}
\arguments{
\item{x}{A numeric matrix of predictors.}

\item{y}{A numeric vector of outcome values.}
}
\description{
This is an internal function that interfaces directly with the Rust
implementation from linfa. The preferred entry point is via tidymodels,
i.e. with:

\if{html}{\out{<div class="sourceCode">}}\preformatted{model_spec <- linear_reg(engine = "linfa")
model <- fit(model_spec, mpg ~ ., mtcars)
}\if{html}{\out{</div>}}
}
\examples{
x <- matrix(rnorm(300), ncol = 3)
y <- rnorm(100)

m <- .linfa_linear_reg(x, y)
m

predict(m, matrix(rnorm(12), ncol = 3))
}
\keyword{internal}
