% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-logistic_reg.R
\name{.linfa_logistic_reg}
\alias{.linfa_logistic_reg}
\title{Logistic regression with linfa}
\usage{
.linfa_logistic_reg(x, y)
}
\arguments{
\item{x}{A numeric matrix of predictors.}

\item{y}{An integer vector of outcome values.}
}
\description{
This is an internal function that interfaces directly with the Rust
implementation from linfa. The preferred entry point is via tidymodels,
i.e. with:

\if{html}{\out{<div class="sourceCode">}}\preformatted{model_spec <- decision_tree(engine = "linfa", mode = "classification")
model <- fit(model_spec, as.factor(vs) ~ ., mtcars)
}\if{html}{\out{</div>}}
}
\examples{
x <- matrix(rnorm(300), ncol = 3)
y <- sample(1:2, size = 100, replace = TRUE)

m <- linfa_logistic_reg(x, y)
m

predict(m, matrix(rnorm(12), ncol = 3))
}
\keyword{internal}
