% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-multinom_reg.R
\name{.linfa_multinom_reg}
\alias{.linfa_multinom_reg}
\title{Multinomial regression with linfa}
\usage{
.linfa_multinom_reg(x, y)
}
\arguments{
\item{x}{A numeric matrix of predictors.}

\item{y}{An integer vector of outcome values.}
}
\description{
This is an internal function that interfaces directly with the Rust
implementation from linfa. The preferred entry point is via tidymodels,
i.e. with:

\if{html}{\out{<div class="sourceCode">}}\preformatted{model_spec <- multinom_reg(engine = "linfa", mode = "classification")
model <- fit(model_spec, as.factor(cyl) ~ ., mtcars)
}\if{html}{\out{</div>}}
}
\examples{
x <- matrix(rnorm(300), ncol = 3)
y <- sample(1:4, size = 100, replace = TRUE)

m <- .linfa_multinom_reg(x, y)
m

predict(m, matrix(rnorm(12), ncol = 3))
}
\keyword{internal}
