% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fit_decision_tree}
\alias{fit_decision_tree}
\alias{predict_decision_tree}
\title{Fit and predict with a decision tree}
\usage{
fit_decision_tree(x, y, n_features, max_depth)

predict_decision_tree(model, x, n_features)
}
\arguments{
\item{x}{A real-valued vector with length \code{y} times some
whole-number multiple \code{n_features}.}

\item{y}{A vector of integers representing classes.}

\item{n_features}{Integer giving the number of features in x.}

\item{max_depth}{Integer.}

\item{model}{Output from \code{fit_decision_tree()}.}
}
\value{
\itemize{
\item For \code{fit_decision_tree()}, an external pointer to a Rust object.
\item For \code{predict_decision_tree()}, an integer vector of class predictions.
}
}
\description{
This function currently only supports classification.
}
\examples{
x <- rnorm(3e5)
y <- sample(1:4, 1e5, replace = TRUE)

mod <- fit_decision_tree(x, y, n_features = 3L, max_depth = 5L)

mod

preds <- predict_decision_tree(mod, x[1:12], 3)

preds

}
