% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fit_linear_reg_linfa}
\alias{fit_linear_reg_linfa}
\title{Fit a linear regression model and return a pointer to it.}
\usage{
fit_linear_reg_linfa()
}
\description{
Fit a linear regression model and return a pointer to it.
}
