% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-logistic_reg.R
\name{linfa_logistic_reg}
\alias{linfa_logistic_reg}
\title{Logistic regression with linfa}
\usage{
linfa_logistic_reg(x, y)
}
\arguments{
\item{x}{A numeric matrix of predictors.}

\item{y}{An integer vector of outcome values.}
}
\description{
Logistic regression with linfa
}
\examples{
x <- matrix(rnorm(300), ncol = 3)
y <- sample(1:2, size = 100, replace = TRUE)

m <- linfa_logistic_reg(x, y)
m

predict(m, matrix(rnorm(12), ncol = 3))
}
