% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-naive_Bayes.R
\name{linfa_naive_Bayes}
\alias{linfa_naive_Bayes}
\title{Naive bayes with linfa}
\usage{
linfa_naive_Bayes(x, y, smoothness = 1e-09)
}
\arguments{
\item{x}{A numeric matrix of predictors.}

\item{y}{An integer vector of outcome values.}

\item{smoothness}{An non-negative number representing the the relative
smoothness of the class boundary. Smaller examples result in model flexible
boundaries and larger values generate class boundaries that are less
adaptable}
}
\description{
Naive bayes with linfa
}
\examples{
x <- matrix(rnorm(300), ncol = 3)
y <- sample(1:4, size = 100, replace = TRUE)

m <- linfa_naive_Bayes(x, y)
m

predict(m, matrix(rnorm(12), ncol = 3))
}
\keyword{internal}
