% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-svm_linear.R
\name{linfa_svm_linear}
\alias{linfa_svm_linear}
\title{Support vector machines with linfa}
\usage{
linfa_svm_linear(x, y, cost = 1)
}
\arguments{
\item{x}{A numeric matrix of predictors.}

\item{y}{A logical vector of outcome values.}

\item{cost}{A positive number for the cost of predicting a sample within
or on the wrong side of the margin}
}
\description{
Support vector machines with linfa
}
\examples{
x <- matrix(rnorm(300), ncol = 3)
y <- sample(c(0L, 1L), size = 100, replace = TRUE)

m <- linfa_svm_linear(x, y)
m

predict(m, matrix(rnorm(12), ncol = 3))
}
\keyword{internal}
