% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{predict_linear_reg_linfa}
\alias{predict_linear_reg_linfa}
\title{Given a pointer to a model and data, return predictions from the model.}
\usage{
predict_linear_reg_linfa(pointer)
}
\description{
Given a pointer to a model and data, return predictions from the model.
}
