% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{rinfa-package}
\alias{rinfa}
\alias{rinfa-package}
\title{rinfa: Rust bindings for tidymodels}
\description{
A 'parsnip' extension that interfaces with the 'Rust' machine learning library 'linfa' to decrease model fit times for common model types.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/rinfa}
  \item Report bugs at \url{https://github.com/simonpcouch/rinfa/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co}

Authors:
\itemize{
  \item Josiah Parry \email{josiah.parry@gmail.com} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID})
}

}
\keyword{internal}
