% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{rinfa-package}
\alias{rinfa}
\alias{rinfa-package}
\title{rinfa: Rust bindings for tidymodels}
\description{
Rust bindings for tidymodels
}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co}

}
